use Double everywhere

+++ src/adp2/interfaces/TSP.java
@@ -14,6 +14,6 @@ public interface TSP {
      * @return the shortest path connecting the waypoints in m or
      *         Path(EmptyList, -1) if m is not an nxn-Matrix or is not symmetric
      */
-    Path minPath(Matrix<Integer> m);
+    Path minPath(Matrix<Double> m);
     
 }
